Update description.md
Hijacking via changing the redirect_uri wasnt explained correctly. The description state that the third-party server should be validating the redirect_uri, but this is incorrect. Its the IdP that validates the redirect_uri. 

If the redirect_uri isnt appropriately restricted, an attacker could change it to their own domain, and send the initial login url to a victim. Once the victim has completed the auth flow, the auth code would be sent to the attacker's server, and the attacker could then exchange it for creds. But at no point would the third-party server be performing any validation in this scenario. It has to be done by the IdP